requirements.txt
streamlit>=1.28.0
browser-use
python-dotenv
openai
playwright
